fix bug in geometric estimation and bootstrap
Geometric mean:
I was removing the first 4 columns in the 'estim_summary_gm' but there must be
kept. I was also removing the count column before conversion in the 'verif_ab'
script.

Bootstrap:
I was removing infinite values after resampling, resulting in different size of
data for each model. But there are no difference in the results.